@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of composer/satis.
  *
@@ -13,38 +15,22 @@ namespace Composer\Satis\Builder;
 
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Base class for Satis Builders.
- *
- * @author James Hautot <james@rezo.net>
- */
 abstract class Builder implements BuilderInterface
 {
     /** @var OutputInterface $output The output Interface. */
     protected $output;
-
     /** @var string $outputDir The directory where to build. */
     protected $outputDir;
-
     /** @var array $config The parameters from ./satis.json. */
     protected $config;
-
     /** @var bool $skipErrors Skips Exceptions if true. */
     protected $skipErrors;
 
-    /**
-     * Base Constructor.
-     *
-     * @param OutputInterface $output     The output Interface
-     * @param string          $outputDir  The directory where to build
-     * @param array           $config     The parameters from ./satis.json
-     * @param bool            $skipErrors Skips Exceptions if true
-     */
-    public function __construct(OutputInterface $output, $outputDir, $config, $skipErrors)
+    public function __construct(OutputInterface $output, string $outputDir, array $config, bool $skipErrors)
     {
         $this->output = $output;
         $this->outputDir = $outputDir;
         $this->config = $config;
-        $this->skipErrors = (bool) $skipErrors;
+        $this->skipErrors = $skipErrors;
     }
 }
