@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of composer/satis.
  *
@@ -13,32 +15,19 @@ namespace Composer\Satis\Builder;
 
 use Composer\Json\JsonFile;
 use Composer\Package\Dumper\ArrayDumper;
+use Composer\Package\PackageInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * Builds the JSON files.
- *
- * @author James Hautot <james@rezo.net>
- */
 class PackagesBuilder extends Builder
 {
     /** @var string packages.json file name. */
     private $filename;
-
     /** @var string included json filename template */
     private $includeFileName;
-
+    /** @var array */
     private $writtenIncludeJsons = [];
 
-    /**
-     * Dedicated Packages Constructor.
-     *
-     * @param OutputInterface $output     The output Interface
-     * @param string          $outputDir  The directory where to build
-     * @param array           $config     The parameters from ./satis.json
-     * @param bool            $skipErrors Escapes Exceptions if true
-     */
-    public function __construct(OutputInterface $output, $outputDir, $config, $skipErrors)
+    public function __construct(OutputInterface $output, string $outputDir, array $config, bool $skipErrors)
     {
         parent::__construct($output, $outputDir, $config, $skipErrors);
 
@@ -47,11 +36,9 @@ class PackagesBuilder extends Builder
     }
 
     /**
-     * Builds the JSON stuff of the repository.
-     *
-     * @param \Composer\Package\PackageInterface[] $packages List of packages to dump
+     * @param PackageInterface[] $packages List of packages to dump
      */
-    public function dump(array $packages)
+    public function dump(array $packages): void
     {
         $packagesByName = [];
         $dumper = new ArrayDumper();
@@ -95,15 +82,7 @@ class PackagesBuilder extends Builder
         $this->pruneIncludeDirectories();
     }
 
-    /**
-     * Find packages replacing the $replaced packages
-     *
-     * @param array $packages
-     * @param string $replaced
-     *
-     * @return array
-     */
-    private function findReplacements($packages, $replaced)
+    private function findReplacements(array $packages, string $replaced): array
     {
         $replacements = [];
         foreach ($packages as $packageName => $packageConfig) {
@@ -118,10 +97,7 @@ class PackagesBuilder extends Builder
         return $replacements;
     }
 
-    /**
-     * Remove all files matching the includeUrl pattern next to just created include jsons
-     */
-    private function pruneIncludeDirectories()
+    private function pruneIncludeDirectories(): void
     {
         $this->output->writeln('<info>Pruning include directories</info>');
         $paths = [];
@@ -181,23 +157,18 @@ class PackagesBuilder extends Builder
         }
     }
 
-    /**
-     * Writes includes JSON Files.
-     *
-     * @param array $packages List of packages to dump
-     * @param string $includesUrl The includes url (optionally containing %hash%)
-     * @param string $hashAlgorithm Hash algorithm {@see hash()}
-     *
-     * @return array The object for includes key in packages.json
-     */
-    private function dumpPackageIncludeJson(array $packages, $includesUrl, $hashAlgorithm = 'sha1')
+    private function dumpPackageIncludeJson(array $packages, string $includesUrl, string $hashAlgorithm = 'sha1'): array
     {
         $filename = str_replace('%hash%', 'prep', $includesUrl);
         $path = $tmpPath = $this->outputDir . '/' . ltrim($filename, '/');
 
         $repoJson = new JsonFile($path);
-        $contents = $repoJson->encode(['packages' => $packages]) . "\n";
+        $options = JSON_UNESCAPED_UNICODE | JSON_UNESCAPED_SLASHES;
+        if ($this->config['pretty-print'] ?? true) {
+            $options |= JSON_PRETTY_PRINT;
+        }
 
+        $contents = $repoJson->encode(['packages' => $packages], $options) . "\n";
         $hash = hash($hashAlgorithm, $contents);
 
         if (false !== strpos($includesUrl, '%hash%')) {
@@ -209,9 +180,10 @@ class PackagesBuilder extends Builder
                 $path = null;
             }
         }
+
         if ($path) {
             $this->writeToFile($path, $contents);
-            $this->output->writeln("<info>wrote packages to $path</info>");
+            $this->output->writeln("<info>Wrote packages to $path</info>");
         }
 
         return [
@@ -220,15 +192,10 @@ class PackagesBuilder extends Builder
     }
 
     /**
-     * Write to a file
-     *
-     * @param string $path
-     * @param string $contents
-     *
      * @throws \UnexpectedValueException
      * @throws \Exception
      */
-    private function writeToFile($path, $contents)
+    private function writeToFile(string $path, string $contents): void
     {
         $dir = dirname($path);
         if (!is_dir($dir)) {
@@ -261,11 +228,9 @@ class PackagesBuilder extends Builder
     }
 
     /**
-     * Writes the packages.json of the repository.
-     *
      * @param array $repo Repository information
      */
-    private function dumpPackagesJson($repo)
+    private function dumpPackagesJson(array $repo): void
     {
         if (isset($this->config['notify-batch'])) {
             $repo['notify-batch'] = $this->config['notify-batch'];
