@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of composer/satis.
  *
@@ -30,9 +32,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-/**
- * @author Jordi Boggiano <j.boggiano@seld.be>
- */
 class BuildCommand extends BaseCommand
 {
     protected function configure()
@@ -45,6 +44,7 @@ class BuildCommand extends BaseCommand
                 new InputArgument('output-dir', InputArgument::OPTIONAL, 'Location where to output built files', null),
                 new InputArgument('packages', InputArgument::IS_ARRAY | InputArgument::OPTIONAL, 'Packages that should be built. If not provided, all packages are built.', null),
                 new InputOption('repository-url', null, InputOption::VALUE_OPTIONAL, 'Only update the repository at given url', null),
+                new InputOption('repository-strict', null, InputOption::VALUE_NONE, 'Also apply the repository filter when resolving dependencies'),
                 new InputOption('no-html-output', null, InputOption::VALUE_NONE, 'Turn off HTML view'),
                 new InputOption('skip-errors', null, InputOption::VALUE_NONE, 'Skip Download or Archive errors'),
                 new InputOption('stats', null, InputOption::VALUE_NONE, 'Display the download progress bar'),
@@ -58,6 +58,7 @@ The json config file accepts the following keys:
 
 - <info>"repositories"</info>: defines which repositories are searched
   for packages.
+- <info>"repositories-dep"</info>: define additional repositories for dependencies
 - <info>"output-dir"</info>: where to output the repository files
   if not provided as an argument when calling build.
 - <info>"require-all"</info>: boolean, if true, all packages present
@@ -74,8 +75,12 @@ The json config file accepts the following keys:
   requirements' dependencies.
 - <info>"require-dev-dependencies"</info>: works like require-dependencies
   but requires dev requirements rather than regular ones.
+- <info>"only-dependencies"</info>: only require dependencies - choose this if you want to build
+  a mirror of your project's dependencies without building packages for the main project repositories.
 - <info>"config"</info>: all config options from composer, see
   http://getcomposer.org/doc/04-schema.md#config
+- <info>"strip-hosts"</info>: boolean or an array of domains, IPs, CIDR notations, '/local' (=localnet and other reserved)
+  or '/private' (=private IPs) to be stripped from the output. If set and non-false, local file paths are removed too.
 - <info>"output-html"</info>: boolean, controls whether the repository
   has an html page as well or not.
 - <info>"name"</info>: for html output, this defines the name of the
@@ -86,6 +91,7 @@ The json config file accepts the following keys:
   building the html output.
 - <info>"abandoned"</info>: Packages that are abandoned. As the key use the
   package name, as the value use true or the replacement package.
+- <info>"blacklist"</info>: Packages and versions which should be excluded from the final package list.
 - <info>"notify-batch"</info>: Allows you to specify a URL that will
   be called every time a user installs a package, see
   https://getcomposer.org/doc/05-repositories.md#notify-batch
@@ -97,16 +103,11 @@ EOT
     }
 
     /**
-     * @param InputInterface  $input  The input instance
-     * @param OutputInterface $output The output instance
-     *
      * @throws JsonValidationException
      * @throws ParsingException
      * @throws \Exception
-     *
-     * @return int
      */
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): int
     {
         $verbose = $input->getOption('verbose');
         $configFile = $input->getArgument('file');
@@ -175,7 +176,7 @@ EOT
         $packageSelection = new PackageSelection($output, $outputDir, $config, $skipErrors);
 
         if (null !== $repositoryUrl) {
-            $packageSelection->setRepositoryFilter($repositoryUrl);
+            $packageSelection->setRepositoryFilter($repositoryUrl, (bool) $input->getOption('repository-strict'));
         } else {
             $packageSelection->setPackagesFilter($packagesFilter);
         }
@@ -188,6 +189,8 @@ EOT
             $downloads->setInput($input);
             $downloads->dump($packages);
         }
+
+        $packages = $packageSelection->clean();
 
         if ($packageSelection->hasFilterForPackages() || $packageSelection->hasRepositoryFilter()) {
             // in case of an active filter we need to load the dumped packages.json and merge the
@@ -213,10 +216,7 @@ EOT
         return 0;
     }
 
-    /**
-     * @return Config
-     */
-    private function getConfiguration()
+    private function getConfiguration(): Config
     {
         $config = new Config();
 
@@ -233,12 +233,7 @@ EOT
         return $config;
     }
 
-    /**
-     * @throws \RuntimeException
-     *
-     * @return string
-     */
-    private function getComposerHome()
+    private function getComposerHome(): string
     {
         $home = getenv('COMPOSER_HOME');
         if (!$home) {
@@ -259,17 +254,10 @@ EOT
     }
 
     /**
-     * Validates the syntax and the schema of the current config json file
-     * according to satis-schema.json rules.
-     *
-     * @param string $configFile The json file to use
-     *
      * @throws ParsingException        if the json file has an invalid syntax
      * @throws JsonValidationException if the json file doesn't match the schema
-     *
-     * @return bool true on success
      */
-    private function check($configFile)
+    private function check(string $configFile): bool
     {
         $content = file_get_contents($configFile);
 
@@ -292,6 +280,7 @@ EOT
                 foreach ((array) $validator->getErrors() as $error) {
                     $errors[] = ($error['property'] ? $error['property'] . ' : ' : '') . $error['message'];
                 }
+
                 throw new JsonValidationException('The json config file does not match the expected JSON schema', $errors);
             }
 
