@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /*
  * This file is part of composer/satis.
  *
@@ -35,7 +37,7 @@ class WebBuilderDumpTest extends TestCase
     /** @var vfsStreamDirectory */
     protected $root;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->rootPackage = new RootPackage('dummy root package', 0, 0);
 
@@ -44,7 +46,7 @@ class WebBuilderDumpTest extends TestCase
         $this->root = $this->setFileSystem();
     }
 
-    protected function setFileSystem()
+    protected function setFileSystem(): vfsStreamDirectory
     {
         vfsStreamWrapper::register();
         $root = vfsStream::newDirectory('build');
@@ -61,8 +63,8 @@ class WebBuilderDumpTest extends TestCase
 
         $html = $this->root->getChild('build/index.html')->getContent();
 
-        $this->assertRegExp('/<title>dummy root package Composer repository<\/title>/', $html);
-        $this->assertRegExp('{<h3 id="[^"]+" class="panel-title package-title">\s*<a href="#vendor/name" class="anchor">\s*<svg[^>]*>.+</svg>\s*vendor/name\s*</a>\s*</h3>}si', $html);
+        $this->assertRegExp('/<title>dummy root package<\/title>/', $html);
+        $this->assertRegExp('{<div id="[^"]+" class="card-header[^"]+">\s*<a href="#vendor/name" class="[^"]+">\s*<svg[^>]*>.+</svg>\s*vendor/name\s*</a>\s*</div>}si', $html);
         $this->assertFalse((bool) preg_match('/<p class="abandoned">/', $html));
     }
 
@@ -75,7 +77,7 @@ class WebBuilderDumpTest extends TestCase
 
         $html = $this->root->getChild('build/index.html')->getContent();
 
-        $this->assertRegExp('/<title>A Composer repository<\/title>/', $html);
+        $this->assertRegExp('/<title>A<\/title>/', $html);
     }
 
     public function testDependencies()
@@ -91,10 +93,7 @@ class WebBuilderDumpTest extends TestCase
         $this->assertRegExp('/<a href="#dummytest">dummytest<\/a>/', $html);
     }
 
-    /**
-     * @return array
-     */
-    public function dataAbandoned()
+    public function dataAbandoned(): array
     {
         $data = [];
 
@@ -117,7 +116,7 @@ class WebBuilderDumpTest extends TestCase
      * @param bool|string $abandoned
      * @param string $expected
      */
-    public function testAbandoned($abandoned, $expected)
+    public function testAbandoned($abandoned, string $expected)
     {
         $webBuilder = new WebBuilder(new NullOutput(), vfsStream::url('build'), [], false);
         $webBuilder->setRootPackage($this->rootPackage);
